Attempt a fast path loading all songs for playlist
Edit: moved to bliss

When generating a playlist profiling shows that part that takes the most
time is loading all the songs from the database for the initial sort.
rusqlite doesn't have a "load everything at once" method unfortunately
and going back and forth between sqlite and rust for each step through
the result set probably isn't particularly fast. Switching to alternate
bindings could be an interesting exercise.

This is an attempt to load the least possible data for the initial sort,
and then go back to the DB once we have a list of neighbours.

The only profiling I've done so far is with the flamegraph crate (and
the time command). My sample size is about 140k tracks. I'm running on
spinning rust.

Motivation:

    $ sqlite3 ~/.local/share/bliss-rs/songs.db 'select song.path, feature, album, track_number, title, artist, genre, version from feature inner join song on song.id = feature.song_id where song.analyzed = true order by path;' | pv > /dev/null
    476MiB 0:00:05 [86.5MiB/s]

    $ sqlite3 ~/.local/share/bliss-rs/songs.db 'select song.id, feature, feature_index from feature inner join song on song.id = feature.song_id where song.analyzed = true;' | pv  > /dev/null
    75.0MiB 0:00:01 [44.5MiB/s]

The good:
* it works and it is indeed faster

The bad:
* duplicates code
* doesn't work with most modes yet (--album-playlist, --seed-song,
  --deduplicate-songs), most of them expect some more stuff to be on the
  Song objects
* some shortcuts and hackishness around error handling and converting
  types because I don't know rust
* need to check how much of a difference just removing the sort from the
  original query helps since it isn't needed (we are sorting based on
  distance anyway)