Make update ignore songs that failed analysis by default
If I have some known errors in my music collection I would like to have
the opportunity to ignore them and proceed with the rest of the
collection and fix the up at a later time.

Here I've chosen to have this as the default behaviour and have a
--force or similar argument later.

I've also added a new query method that only queries paths, since that
was all that was needed, making the query a little faster.

I wish I could get rid of that procedural `for result in results { ...
}` block but I don't know how to handle the potential ResqliteErrors in
a streaming fashion. I thought thats what map_err was supposed to do...

Note that this means it'll no longer be warning users of outdated
feature versions in this code path. I would prefer to move that to a
standalone check that was done earlier on in the program common to more
code paths anyway.