Remove attempted fast path for loading all songs from db
The goal of this fast path was to only load the features table into RAM
for the initial sorting of the entire dataset. Then only load the full
songs for those that made it into the top of the list. This has the
benefit that we aren't loading all the extra information that we don't
need (like path, title, etc) into rust from sqlite for the entire music
library when we don't need it (we only need the features to compute
distances).

Unfortunately it's become difficult to implement that fast path in the
bliss library crate now that that the SQL queries have been moved into a
single inflexible interface. See the commit message in bliss-rs for
more.

Anyway, I'll do a little more testing but it only seems to save a few
hundred ms in real world testing, so maybe I won't bother
re-implementing it.