Make euclidean the default again, don't add seed songs back to playlist
I believe that the euclidean sort method shows significantly better
results than the extended isolation forest, particularly for small seed
song sets. For just one of two songs the isolation forest seems to just
return the first songs in the DB by song_id, for larger seed sets it's
output isn't stable and there are some wild picks in the resulting
playlists.

The prior implementation would clear the queue out before adding the new
tracks. I think this may still be useful for the from-current-song case,
but more importantly the sort() output includes the seed songs so since
the current implementation isn't clearing them is always adds them as
duplicates to the playlist. Change it back to only adding the new songs
to the playlist.